before midterm present

+++ views/product/view.blade.php
@@ -1,7 +1,10 @@
 @extends('layout.newDefault')
 @section('content')
 @include('layout.newNav')
-@include('layout.menu_admin')
+
+@if(Entrust::hasRole('Admin'))
+	@include('layout.menu_admin')
+@endif
 <div style="margin-left:100px">
 	<br/>
 	<h1>{{ $product -> name }}</h1>
@@ -13,11 +16,13 @@
 	<p>{{ $product -> sex }}</p>
 	<p>{{ $product -> quantity }}</p>
 
-<a href="{{ URL::route('product-edit-view',array($product -> id)) }}">Edit Product</a>
-{{ Form::open(array('url' => 'products/delete','method' => 'delete')) }}
-{{ Form::token() }}
-{{ Form::hidden('id',$product->id) }}
-{{ Form::submit('Delete') }}
-{{ Form::close() }}
-<br>
+@if(Entrust::hasRole('Admin'))
+	<a href="{{ URL::route('product-edit-view',array($product -> id)) }}">Edit Product</a>
+	{{ Form::open(array('url' => 'products/delete','method' => 'delete')) }}
+	{{ Form::token() }}
+	{{ Form::hidden('id',$product->id) }}
+	{{ Form::submit('Delete') }}
+	{{ Form::close() }}
+	<br>
+@endif
 @endsection
